sort/requirements.txt
streamlit
streamlit-lottie
numpy
pandas
Pillow
pyrebase4
requests
opencv-python-headless